Sync'ed context to models for tracer
Stored the url in the model and then parsed it back out and set the state context directly. Can't navigateLink(url) because then that async loads the data which don't want because the models already supplied by tracer.
Not ideal because not resetting the full context (crumbs missing) but it's ok for now